feat(SignIn): User SignIn feature added Generating JWT done

+++ pkg/domain/user.go
@@ -1,7 +1,7 @@
 package domain
 
 type User struct {
-	Id         uint   `json:"id" gorm:"primaryKey;autoIncrement:true;unique"`
+	Id         uint64 `json:"id" gorm:"primaryKey;autoIncrement:true;unique"`
 	Firstname  string `json:"firstname" validate:"required,min=4,max=16"`
 	Lastname   string `json:"lastname" validate:"required,min=4,max=16"`
 	Password   string `json:"password" validate:"required,min=6,max=16"`
